pyproject.toml
[project]
name = "cowardly-irregular-sprite-gen"
version = "0.1.0"
description = "AI sprite generation pipeline for Cowardly Irregular"
requires-python = ">=3.11,<3.14"
dependencies = [
    "torch>=2.1.0",
    "diffusers>=0.25.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "safetensors>=0.4.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "sentencepiece>=0.1.99",
    "controlnet-aux>=0.0.7",
]

[project.optional-dependencies]
pixeloe = ["pixeloe>=0.1.0"]

[tool.uv]
find-links = ["https://download.pytorch.org/whl/cu124"]
